% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop_times_spjp.R
\name{build_daytype_map_df}
\alias{build_daytype_map_df}
\title{Build a data.frame mapping DayType to UicOperatingPeriods}
\usage{
build_daytype_map_df(x, ns)
}
\arguments{
\item{x}{NeTEx XML document}

\item{ns}{Namespace vector}
}
\value{
data.frame with columns: daytype_id, uic_id, from, to
}
\description{
Build a data.frame mapping DayType to UicOperatingPeriods
}
