% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spjp.R
\name{build_spjp_df}
\alias{build_spjp_df}
\title{Stop point information for all scheduled stop points}
\usage{
build_spjp_df(x, ns = xml_ns(x))
}
\arguments{
\item{x}{NeTEx XML file}

\item{ns}{Namespace}
}
\value{
data.frame
}
\description{
Stop point information for all scheduled stop points
}
\examples{
bus_371 <- xml2::read_xml(netexr_example("NX-PI-01_DE_NAP_LINE_mvv_d-REGBU-85707123_20251117.xml"))

# One stop point
spjp <- xml2::xml_find_first(bus_371, ".//d1:StopPointInJourneyPattern",
                             ns = xml2::xml_ns(bus_371))
xml2::xml_name(spjp)
extract_one_spjp(spjp)

# All stop points
spjp_df <- build_spjp_df(bus_371)
head(spjp_df)
}
