% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop.R
\name{build_stop_df}
\alias{build_stop_df}
\title{Stop point information for all scheduled stop points}
\usage{
build_stop_df(x, ns = xml_ns(x))
}
\arguments{
\item{x}{NeTEx XML file}

\item{ns}{Namespace}
}
\value{
data.frame
}
\description{
Stop point information for all scheduled stop points
}
\examples{
bus_371 <- xml2::read_xml(netexr_example("NX-PI-01_DE_NAP_LINE_mvv_d-REGBU-85707123_20251117.xml"))
stop_df <- build_stop_df(bus_371)

}
