% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line.R
\name{build_times_df}
\alias{build_times_df}
\title{Build times_df for all journeys}
\usage{
build_times_df(journeys, line_name, line_type, ns)
}
\arguments{
\item{journeys}{Journeys}

\item{line_name}{Line Name}

\item{line_type}{Line Type}

\item{ns}{Namespace}
}
\value{
data.frame
}
\description{
Build times_df for all journeys
}
\examples{
bus_371 <- xml2::read_xml(netexr_example("NX-PI-01_DE_NAP_LINE_mvv_d-REGBU-85707123_20251117.xml"))
ns <- xml_ns(bus_371)

# 1. Get the line info (since there is only one line)
line_info <- get_line_info(bus_371, ns)

# 2. Get all journeys
journeys <- xml_find_all(bus_371, ".//d1:ServiceJourney", ns)

# 3. Build the times dataframe
times_df <- build_times_df(journeys, line_info$line, line_info$line_type, ns)
}
