% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop_times_spjp.R
\name{date_in_valid_bits}
\alias{date_in_valid_bits}
\title{Check if a date is active in a ValidDayBits string}
\usage{
date_in_valid_bits(date, from_date, valid_bits)
}
\arguments{
\item{date}{Date to check}

\item{from_date}{FromDate of UicOperatingPeriod (Date)}

\item{valid_bits}{string of 0/1}
}
\value{
TRUE if date is active, FALSE otherwise
}
\description{
Check if a date is active in a ValidDayBits string
}
