% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line.R
\name{extract_one_passing}
\alias{extract_one_passing}
\title{Extract one passing-time node}
\usage{
extract_one_passing(p_node, line_name, line_type, ns)
}
\arguments{
\item{p_node}{Passing Node}

\item{line_name}{Line Name}

\item{line_type}{Line Type}

\item{ns}{Namespace}
}
\value{
data.frame
}
\description{
Extract one passing-time node
}
