% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spjp.R
\name{extract_one_spjp}
\alias{extract_one_spjp}
\title{Extract one stop point StopPointInJourneyPattern}
\usage{
extract_one_spjp(spjp_node, ns = xml_ns(spjp_node))
}
\arguments{
\item{spjp_node}{StopPointInJourneyPattern xml node}

\item{ns}{Namespace}
}
\value{
data.frame
}
\description{
Extract information about the StopPointInJourneyPattern (`spjp_id` and `scheduled_stop_id`)
}
\examples{
bus_371 <- xml2::read_xml(netexr_example("NX-PI-01_DE_NAP_LINE_mvv_d-REGBU-85707123_20251117.xml"))

spjp <- xml2::xml_find_first(bus_371, ".//d1:StopPointInJourneyPattern",
                             ns = xml2::xml_ns(bus_371))

xml2::xml_name(spjp)

extract_one_spjp(spjp)

spjp_df <- build_spjp_df(bus_371)
head(spjp_df)

}
