% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop.R
\name{extract_one_stop}
\alias{extract_one_stop}
\title{Extract one stop point}
\usage{
extract_one_stop(s_node, ns = xml_ns(s_node))
}
\arguments{
\item{s_node}{ScheduledStopPoint XML node}

\item{ns}{Namespace}
}
\value{
data.frame
}
\description{
Extract one stop point
}
\examples{
bus_371 <- xml2::read_xml(netexr_example("NX-PI-01_DE_NAP_LINE_mvv_d-REGBU-85707123_20251117.xml"))

ssp <- xml2::xml_find_all(
  bus_371,
  './/d1:ScheduledStopPoint[d1:Name = "Baiernrain"]',
  ns = xml2::xml_ns(bus_371)
)

xml2::xml_name(ssp[[1]])

extract_one_stop(ssp[[1]])
extract_one_stop(ssp[[2]])

stop_df <- build_stop_df(bus_371)
head(stop_df)

}
