% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop_times_spjp.R
\name{filter_journeys_by_date}
\alias{filter_journeys_by_date}
\title{Filter ServiceJourneys by target date}
\usage{
filter_journeys_by_date(journeys, daytype_map_df, target_date, ns)
}
\arguments{
\item{journeys}{list of ServiceJourney nodes}

\item{daytype_map_df}{Data.frame from build_daytype_map_df()}

\item{target_date}{Date object}

\item{ns}{Namespace vector}
}
\value{
filtered list of journeys
}
\description{
Filter ServiceJourneys by target date
}
