% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line.R
\name{get_line_info}
\alias{get_line_info}
\title{Get the line information from a NeTEx file}
\usage{
get_line_info(x, ns = xml_ns(x))
}
\arguments{
\item{x}{NeTEx XML file}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by xml_ns().}
}
\value{
a list of `line` and `line_type`.
}
\description{
This function retrieves the Line, Name and TransportMode.
It works under the assumption that the `x` only contains information on one
line.
}
\examples{
bus_371 <- xml2::read_xml(netexr_example("NX-PI-01_DE_NAP_LINE_mvv_d-REGBU-85707123_20251117.xml"))
get_line_info(bus_371)
}
