% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop_times_spjp.R
\name{get_stop_times_spjp}
\alias{get_stop_times_spjp}
\title{Get stop times for a specific StopPointInJourneyPattern (modular & data.frame calendar)}
\usage{
get_stop_times_spjp(x, spjp_id, target_date = NULL)
}
\arguments{
\item{x}{NeTEx XML document}

\item{spjp_id}{StopPointInJourneyPattern ID}

\item{target_date}{Optional date (YYYY-MM-DD)}
}
\value{
data.frame with columns: spjp_id, stop_time, line, line_type
}
\description{
Get stop times for a specific StopPointInJourneyPattern (modular & data.frame calendar)
}
