% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gettext_ns}
\alias{gettext_ns}
\title{Safe text extraction from a NeTEx node}
\usage{
gettext_ns(node, xpath, ns = xml_ns(x))
}
\arguments{
\item{node}{NeTEx node}

\item{xpath}{A string containing an xpath (1.0) expression.}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by xml_ns().}
}
\value{
A character vector, the same length as x.
}
\description{
Safe text extraction from a NeTEx node
}
