% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop_times_spjp.R
\name{journey_operates_on_date}
\alias{journey_operates_on_date}
\title{Check if a ServiceJourney operates on a given date}
\usage{
journey_operates_on_date(sj, daytype_map_df, target_date, ns)
}
\arguments{
\item{sj}{ServiceJourney xml_node}

\item{daytype_map_df}{Data.frame from build_daytype_map_df()}

\item{target_date}{Date object}

\item{ns}{Namespace vector}
}
\value{
TRUE if the journey operates on target_date, FALSE otherwise
}
\description{
Check if a ServiceJourney operates on a given date
}
