% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{netexr_example}
\alias{netexr_example}
\title{Get path to a NeTExR example}
\usage{
netexr_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
NeTExR comes bundled with a number of sample files in its ‘inst/extdata’
directory. This function makes them easy to access.
}
