% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line.R
\name{process_one_journey}
\alias{process_one_journey}
\title{Process one ServiceJourney}
\usage{
process_one_journey(journey, line_name, line_type, ns)
}
\arguments{
\item{journey}{Journey}

\item{line_name}{Line Name}

\item{line_type}{Line Type}

\item{ns}{Namespace}
}
\value{
data.frame
}
\description{
Process one ServiceJourney
}
